Move food feed inside tabs
